pyproject.toml
[tool.poetry]
name = "shopback"
version = "0.1.0"
description = ""
authors = ["Bolat Akhmadiyev <bolat@akhmadiyev.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.3.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
